+++ lib/src/commands/do/list/deps.dart
@@ -10,9 +10,9 @@ import 'dart:io';
 import 'package:args/command_runner.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:path/path.dart' as path;
-import '../../backend/add_repository_helper.dart';
-import '../../backend/constants.dart';
-import '../../backend/repo_folder_resolver.dart';
+import '../../../backend/add_repository_helper.dart';
+import '../../../backend/constants.dart';
+import '../../../backend/repo_folder_resolver.dart';
 
 /// Command to list dependencies of a project from the master workspace.
 class ListDepsCommand extends Command<dynamic> {