+++ lib/src/commands/do/list/tickets.dart
@@ -10,8 +10,8 @@ import 'package:args/command_runner.dart';
 import 'package:gg_console_colors/gg_console_colors.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:path/path.dart' as path;
-import '../../backend/constants.dart';
-import '../../backend/workspace_utils.dart';
+import '../../../backend/constants.dart';
+import '../../../backend/workspace_utils.dart';
 
 /// Command to list all tickets and their descriptions.
 class ListTicketsCommand extends Command<void> {