+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - refactor(gg_multi): replace do cancel-review with do review --abort
+- refactor(gg_multi): move ls under do
 
 ### Removed